Start refactoring of simulation
